+++ Delta.CertXplorer/src/Delta.SmartCard/Properties/AssemblyInfo.cs
@@ -1,36 +1,19 @@
 ﻿using System.Reflection;
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-// General Information about an assembly is controlled through the following 
-// set of attributes. Change these attribute values to modify the information
-// associated with an assembly.
-[assembly: AssemblyTitle("Delta.SmartCardLib")]
-[assembly: AssemblyDescription("")]
+[assembly: AssemblyTitle("Delta.SmartCard")]
+[assembly: AssemblyDescription("PC/SC Smart card readers handling")]
 [assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("Delta.SmartCardLib")]
-[assembly: AssemblyCopyright("Copyright ©  2014")]
+[assembly: AssemblyCompany("Delta Software")]
+[assembly: AssemblyProduct("Delta.SmartCard")]
+[assembly: AssemblyCopyright("")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
-
-// Setting ComVisible to false makes the types in this assembly not visible 
-// to COM components.  If you need to access a type in this assembly from 
-// COM, set the ComVisible attribute to true on that type.
 [assembly: ComVisible(false)]
+[assembly: AssemblyVersion(AssemblyInfo.Version)]
+[assembly: AssemblyFileVersion(AssemblyInfo.Version)]
 
-// The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("7ce47cd4-d3c5-4aa2-8448-b51c728f5548")]
-
-// Version information for an assembly consists of the following four values:
-//
-//      Major Version
-//      Minor Version 
-//      Build Number
-//      Revision
-//
-// You can specify all the values or you can default the Build and Revision Numbers 
-// by using the '*' as shown below:
-// [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("1.0.0.0")]
-[assembly: AssemblyFileVersion("1.0.0.0")]
+internal static class AssemblyInfo
+{
+    public const string Version = "0.1.0.0";
+}